+++ src/index.tsx
@@ -4,6 +4,7 @@ import './index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import { BrowserRouter, Switch, Route } from 'react-router-dom'
+import { UserProvider } from './context'
 import Dashboard from './components/DashBoard/DashBoard'
 import Users from './components/Users/Users'
 import Cards from './components/Cards/Cards'
@@ -11,23 +12,27 @@ import Audits from './components/Audits/Audits'
 
 ReactDOM.render(
   <BrowserRouter>
-    <Switch>
-      <Route exact path="/">
-        <App />
-      </Route>
-      <Route exact path="/dashboard">
-        <Dashboard />
-      </Route>
-      <Route path="/users">
-        <Users />
-      </Route>
-      <Route path="/cards">
-        <Cards />
-      </Route>
-      <Route path="/audits">
-        <Audits />
-      </Route>
-    </Switch>
+    
+      <Switch>
+        <Route exact path="/">
+        <UserProvider>
+          <App />
+        </UserProvider>
+        </Route>
+        <Route exact path="/dashboard">
+          <Dashboard />
+        </Route>
+        <Route path="/users">
+          <Users />
+        </Route>
+        <Route path="/cards">
+          <Cards />
+        </Route>
+        <Route path="/audits">
+          <Audits />
+        </Route>
+      </Switch>
+    
   </BrowserRouter>,
   document.getElementById('root')
 );
